+++ src/components/Header/index.tsx
@@ -1,13 +1,9 @@
-import React, { useContext } from 'react';
 import styled from '@emotion/styled';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
-import { useRouter } from 'next/router';
-import Menu from './Menu';
-import { UserContext } from 'UserContext';
-import UserAvatar from 'components/UserAvatar';
+import Link from 'components/Link';
+import UserMenu from './UserMenu';
 
 const Header = styled.header`
   flex-grow: 1;
@@ -20,33 +16,16 @@ const Title = styled(Typography)`
 `;
 
 export default function ButtonAppBar() {
-  const router = useRouter();
-
-  const { currentUser } = useContext(UserContext);
-
-  let loginButton;
-  if (currentUser) {
-    const { url } = currentUser;
-    loginButton = <UserAvatar url={url} />;
-  } else {
-    loginButton = (
-      <Button color="inherit" onClick={() => router.push('/login')}>
-        Login
-      </Button>
-    );
-  }
-
   return (
     <Header>
       <AppBar position="static" color="transparent">
         <Toolbar>
-          <Menu user={currentUser} />
           <Title variant="h6">
-            <span onClick={() => router.push('/')}>Keiba Post</span>
+            <Link href="/">Keiba Post</Link>
           </Title>
-          {loginButton}
+          <UserMenu />
         </Toolbar>
       </AppBar>
-    </Header>
+    </Header >
   );
 }
