+++ src/components/Link/index.tsx
@@ -1,0 +1,15 @@
+import { ReactNode } from 'react';
+import NextLink from 'next/link';
+import Link from '@mui/material/Link';
+
+type Props = { href: string; prefetch?: boolean; children: ReactNode };
+
+export default function LinkComponent({ href, prefetch = false, children }: Props) {
+  return (
+    <NextLink href={href} prefetch={prefetch}>
+      <Link href={href} color="inherit" underline="none">
+        {children}
+      </Link>
+    </NextLink>
+  );
+}
